requirements.txt
tensorflow>=2.10.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0
torch>=1.12.0
torchvision>=0.13.0
openpyxl>=3.0.0
streamlit>=1.28.0
opencv-python>=4.8.0 